+++ zcash_primitives/src/lib.rs
@@ -30,29 +30,3 @@ pub(crate) mod encoding;
 pub mod extensions;
 pub mod merkle_tree;
 pub mod transaction;
-
-#[deprecated(note = "This module is deprecated; use `::zcash_protocol::consensus` instead.")]
-pub mod consensus {
-    pub use zcash_protocol::consensus::*;
-}
-#[deprecated(note = "This module is deprecated; use `::zcash_protocol::constants` instead.")]
-pub mod constants {
-    pub use zcash_protocol::constants::*;
-}
-#[deprecated(note = "This module is deprecated; use `::zcash_protocol::memo` instead.")]
-pub mod memo {
-    pub use zcash_protocol::memo::*;
-}
-#[deprecated(note = "This module is deprecated; use the `zip32` crate instead.")]
-pub mod zip32 {
-    pub use zip32::*;
-}
-#[deprecated(note = "This module is deprecated; use the `zcash_transparent` crate instead.")]
-pub mod legacy {
-    pub use transparent::address::*;
-    #[cfg(feature = "transparent-inputs")]
-    #[deprecated(note = "This module is deprecated; use `::zcash_transparent::keys` instead.")]
-    pub mod keys {
-        pub use transparent::keys::*;
-    }
-}